lop: split receive-pack logic into module
