+++ Documentation/howto/maintain-git.adoc
@@ -593,41 +593,3 @@ to move the fix to apply to the other branch, perhaps like this:
       $ mf=refs/merge-fix
       $ git update-ref $mf/$the_other_branch $mf/ai/topic
       $ git update-ref -d $mf/ai/topic
-
-Coverage builds with Clang/LLVM
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-Clang does not ship `libgcov`, so a straight `make coverage-compile`
-fails on macOS and other LLVM based environments when the linker
-cannot satisfy `-lgcov`.  The coverage targets now probe the compiler
-for an appropriate runtime and fall back to Clang's
-`libclang_rt.profile*.a` archives automatically.  To gather coverage
-data with the system toolchain on macOS, run:
-
-----
-$ make GCOV="llvm-cov gcov" coverage-compile
-$ (cd t && GIT_TEST_LOP_COVERAGE=1 GCOV="llvm-cov gcov" ./t5711-lop-offload.sh)
-$ make GCOV="llvm-cov gcov" coverage-report
-----
-
-If you use a non-standard toolchain layout you can still override the
-detected runtime explicitly by passing `COVERAGE_LIBS=/path/to/archive`
-on the command line.
-
-Generating the optional HTML summary uses `gcov2perl` from the Perl
-`Devel::Cover` package by default.  Install it with your preferred tool
-(for example `brew install cpanminus && cpanm Devel::Cover` or `cpan
-Devel::Cover`), or tell the build system where to find the binary:
-
-----
-$ make GCOV2PERL=/opt/homebrew/bin/gcov2perl cover_db_html
-----
-
-You can alternatively rely on `gcovr` when Perl is unavailable.  Set the
-HTML backend to `gcovr` (and override the binary path if needed) to
-render an HTML report in `cover_db_html/index.html` without the
-`gcov2perl` and `cover` tools:
-
-----
-$ make COVERAGE_HTML=gcovr GCOVR=/opt/homebrew/bin/gcovr coverage-html
-----
