+++ t/t5710-promisor-remote-capability.sh
@@ -4,12 +4,6 @@ test_description='handling of promisor remote advertisement'
 
 . ./test-lib.sh
 
-if test -n "$GIT_TEST_LOP_COVERAGE"
-then
-        . "$TEST_DIRECTORY"/lib-lop-gcov.sh
-        lop_gcov_prepare
-fi
-
 if ! test_have_prereq PERL_TEST_HELPERS
 then
 	skip_all='skipping promisor remote capabilities tests; Perl not available'
@@ -366,116 +360,16 @@ test_expect_success "clone with promisor.checkFields" '
         check_missing_objects server 1 "$oid"
 '
 
-lop_copy_for_coverage () {
-        base=$1 &&
-        rm -rf "${base}" "${base}-lop" "${base}-client" &&
-        cp -r server "${base}" &&
-        cp -r lop "${base}-lop" &&
-        git -C "${base}" config promisor.advertise true &&
-        git -C "${base}" config promisor.sendFields partialCloneFilter &&
-        git -C "${base}" config remote.lop.url "file://$(pwd)/${base}-lop" &&
-        git -C "${base}-lop" config uploadpack.allowFilter true &&
-        git -C "${base}-lop" config uploadpack.allowAnySHA1InWant true
-}
 
-test_expect_success LOP_GCOV "coverage: promisor filter branch (no acceptance)" '
-        lop_copy_for_coverage server-cover &&
-        test_when_finished "rm -rf server-cover server-cover-lop client-noaccept" &&
-        GIT_TEST_LOP_FORCE_NO_ACCEPTANCE=1 GIT_NO_LAZY_FETCH=0 \
-                git clone -c promisor.acceptfromserver=All \
-                        "file://$(pwd)/server-cover" client-noaccept
-'
 
-test_expect_success LOP_GCOV "coverage: promisor filter branch (empty list)" '
-        lop_copy_for_coverage server-empty &&
-        test_when_finished "rm -rf server-empty server-empty-lop client-empty" &&
-        GIT_TEST_LOP_FORCE_EMPTY_FILTER=1 GIT_NO_LAZY_FETCH=0 \
-                git clone -c promisor.acceptfromserver=All \
-                        "file://$(pwd)/server-empty" client-empty
-'
 
-test_expect_success LOP_GCOV "coverage: promisor filter branch (mismatch)" '
-        lop_copy_for_coverage server-mismatch &&
-        test_when_finished "rm -rf server-mismatch server-mismatch-lop client-mismatch" &&
-        GIT_TEST_LOP_FORCE_MISMATCH_FILTER=1 GIT_NO_LAZY_FETCH=0 \
-                git clone -c promisor.acceptfromserver=All \
-                        "file://$(pwd)/server-mismatch" client-mismatch
-'
 
-test_expect_success LOP_GCOV "coverage: promisor token field acceptance" '
-        lop_copy_for_coverage server-token &&
-        git -C server-token config promisor.sendFields "partialCloneFilter,token" &&
-        git -C server-token config promisor.checkFields "partialCloneFilter,token" &&
-        git -C server-token config remote.lop.token coverage-token &&
-        test_when_finished "rm -rf server-token server-token-lop client-token" &&
-        GIT_NO_LAZY_FETCH=0 git clone -c promisor.acceptfromserver=KnownName \
-                "file://$(pwd)/server-token" client-token &&
-        git -C client-token config --get remote.origin.partialclonefilter >/dev/null
-'
 
-test_expect_success LOP_GCOV "coverage: promisor token url check" '
-        lop_copy_for_coverage server-token-url &&
-        git -C server-token-url config promisor.sendFields "partialCloneFilter,token" &&
-        git -C server-token-url config promisor.checkFields "partialCloneFilter,token" &&
-        git -C server-token-url config remote.lop.token coverage-token &&
-        test_when_finished "rm -rf server-token-url server-token-url-lop client-token-url" &&
-        GIT_NO_LAZY_FETCH=0 git clone -c promisor.acceptfromserver=KnownUrl \
-                "file://$(pwd)/server-token-url" client-token-url
-'
 
-test_expect_success LOP_GCOV "coverage: promisor capability invalid element" '
-        lop_copy_for_coverage server-invalid &&
-        test_when_finished "rm -rf server-invalid server-invalid-lop client-invalid" &&
-        GIT_TEST_LOP_FORCE_INVALID_CAPABILITY=1 GIT_NO_LAZY_FETCH=0 \
-                git clone -c promisor.acceptfromserver=All \
-                        "file://$(pwd)/server-invalid" client-invalid
-'
 
-test_expect_success LOP_GCOV "coverage: promisor capability missing url" '
-        lop_copy_for_coverage server-missing-url &&
-        test_when_finished "rm -rf server-missing-url server-missing-url-lop client-missing-url" &&
-        GIT_TEST_LOP_FORCE_MISSING_URL=1 GIT_NO_LAZY_FETCH=0 \
-                git clone -c promisor.acceptfromserver=All \
-                        "file://$(pwd)/server-missing-url" client-missing-url
-'
 
-test_expect_success LOP_GCOV "coverage: promisor check fields known name" '
-        lop_copy_for_coverage server-knowncheck &&
-        git -C server-knowncheck config promisor.sendFields "partialCloneFilter,token" &&
-        git -C server-knowncheck config remote.lop.token coverage-token &&
-        test_when_finished "rm -rf server-knowncheck server-knowncheck-lop client-knowncheck" &&
-        mkdir client-knowncheck &&
-        git -C client-knowncheck init &&
-        git -C client-knowncheck config promisor.acceptfromserver KnownName &&
-        git -C client-knowncheck config promisor.checkFields token &&
-        git -C client-knowncheck config remote.lop.promisor true &&
-        git -C client-knowncheck config remote.lop.fetch "+refs/heads/*:refs/remotes/lop/*" &&
-        git -C client-knowncheck config remote.lop.url "file://$(pwd)/server-knowncheck-lop" &&
-        git -C client-knowncheck config remote.lop.token coverage-token &&
-        git -C client-knowncheck config remote.origin.url "file://$(pwd)/server-knowncheck" &&
-        git -C client-knowncheck config remote.origin.fetch "+refs/heads/*:refs/remotes/origin/*" &&
-        GIT_NO_LAZY_FETCH=0 git -C client-knowncheck fetch --filter="blob:limit=5k" origin
-'
 
-test_expect_success LOP_GCOV "coverage: promisor known-url mismatch warns" '
-        lop_copy_for_coverage server-urlwarn &&
-        test_when_finished "rm -rf server-urlwarn server-urlwarn-lop client-urlwarn" &&
-        err=$PWD/urlwarn.err &&
-        test_when_finished "rm -f $err" &&
-        env GIT_NO_LAZY_FETCH=0 git \
-                -c promisor.acceptfromserver=KnownUrl \
-                -c remote.lop.url="https://invalid.invalid/lop" \
-                clone "file://$(pwd)/server-urlwarn" client-urlwarn \
-                2>"$err"
-'
 
-test_expect_success LOP_GCOV "coverage: promisor known-url empty advert" '
-        lop_copy_for_coverage server-emptyenv &&
-        test_when_finished "rm -rf server-emptyenv server-emptyenv-lop client-emptyenv" &&
-        env GIT_TEST_LOP_FORCE_EMPTY_URL=1 GIT_NO_LAZY_FETCH=0 \
-                git clone -c promisor.acceptfromserver=KnownUrl \
-                        "file://$(pwd)/server-emptyenv" client-emptyenv
-'
 
 test_expect_success "clone with promisor.advertise set to 'true' but don't delete the client" '
         git -C server config promisor.advertise true &&
@@ -558,21 +452,5 @@ test_expect_success "subsequent fetch from a client when promisor.advertise is f
         check_missing_objects server 1 "$oid"
 '
 
-test_expect_success LOP_GCOV 'coverage: promisor advertisement helpers executed' '
-        lop_assert_gcov_functions promisor-remote.c \
-                reset_advertised_filters \
-                record_advertised_filter \
-                promisor_remote_advertised_filter \
-                parse_one_advertised_remote \
-                all_fields_match \
-                should_accept_remote &&
-        lop_assert_gcov_function_coverage promisor-remote.c 85 \
-                reset_advertised_filters \
-                record_advertised_filter \
-                promisor_remote_advertised_filter \
-                parse_one_advertised_remote \
-                all_fields_match \
-                should_accept_remote
-'
 
 test_done
